rename RenderProgress to render-progress for consistency
